feat(darwin): manually parse battery info
resolves #2139
